#987 don't wrap org.openqa.selenium.InvalidArgumentException
